# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Payment Relay API
  version: 0.3.0
  description: |
    Internal payment relay — systems, wallets, payouts, invoices, reports, and pay page.

    **Auth**: Protected routes require `X-API-Key: sk_live_...` header.

    **Amounts**: All monetary values are in minor units (cents).
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /invoices:
    post:
      tags: [invoices]
      summary: Create invoice with QR code
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInvoiceRequest'
      responses:
        '200':
          description: Invoice created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceResponse'
    get:
      tags: [invoices]
      summary: List invoices
      security:
        - ApiKeyAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [open, paid, expired, cancelled]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InvoiceResponse'
...
  /reports/invoices:
    get:
      tags: [reports]
      summary: Invoice report
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/ReportFrom'
        - $ref: '#/components/parameters/ReportTo'
        - $ref: '#/components/parameters/ReportFormat'
        - name: status
          in: query
          schema:
            type: string
      responses:
        '200':
          description: JSON summary or CSV export
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: "System API key from POST /systems (format: sk_live_...)"
    AdminTokenAuth:
      type: apiKey
      in: header
      name: X-Admin-Token
      description: Session token from POST /admin/login (Bearer also accepted)
  parameters:
...
    ReportFrom:
      name: from
      in: query
      required: true
      schema:
        type: string
        format: date-time
    ReportTo:
      name: to
      in: query
      required: true
      schema:
        type: string
        format: date-time
    ReportFormat:
      name: format
      in: query
      schema:
        type: string
        enum: [json, csv]
        default: json
...
  schemas:
...
    CreateInvoiceRequest:
      type: object
      required: [amount, currency, country]
      properties:
        amount:
          type: integer
          format: int64
        currency:
          type: string
        country:
          type: string
        description:
          type: string
        expires_in_hours:
          type: integer
          default: 24
    InvoiceResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        reference:
          type: string
        system_id:
          type: string
          format: uuid
        amount:
          type: integer
        currency:
          type: string
        country:
          type: string
        status:
          type: string
          enum: [open, paid, expired, cancelled]
        description:
          type: string
        expires_at:
          type: string
          format: date-time
        paid_at:
          type: string
          format: date-time
          nullable: true
        transaction_id:
          type: string
          format: uuid
          nullable: true
        qr_url:
          type: string
        qr_code_png_base64:
          type: string
...